+++ src/scrypt.cc
@@ -46,12 +46,12 @@ NAN_METHOD(BScrypt::Derive) {
     return Nan::ThrowTypeError("Sixth argument must be a number.");
 
   const uint8_t *pass = (const uint8_t *)node::Buffer::Data(pbuf);
-  uint32_t passlen = (uint32_t)node::Buffer::Length(pbuf);
+  size_t passlen = (size_t)node::Buffer::Length(pbuf);
   const uint8_t *salt = (const uint8_t *)node::Buffer::Data(sbuf);
   size_t saltlen = (size_t)node::Buffer::Length(sbuf);
   uint64_t N = (uint64_t)Nan::To<int64_t>(info[2]).FromJust();
-  uint64_t r = (uint64_t)Nan::To<int64_t>(info[3]).FromJust();
-  uint64_t p = (uint64_t)Nan::To<int64_t>(info[4]).FromJust();
+  uint32_t r = (uint32_t)Nan::To<int64_t>(info[3]).FromJust();
+  uint32_t p = (uint32_t)Nan::To<int64_t>(info[4]).FromJust();
   size_t keylen = (size_t)Nan::To<int64_t>(info[5]).FromJust();
 
   uint8_t *key = (uint8_t *)malloc(keylen);
@@ -100,12 +100,12 @@ NAN_METHOD(BScrypt::DeriveAsync) {
   v8::Local<v8::Function> callback = info[6].As<v8::Function>();
 
   const uint8_t *pass = (const uint8_t *)node::Buffer::Data(pbuf);
-  uint32_t passlen = (uint32_t)node::Buffer::Length(pbuf);
+  size_t passlen = (size_t)node::Buffer::Length(pbuf);
   const uint8_t *salt = (const uint8_t *)node::Buffer::Data(sbuf);
   size_t saltlen = (size_t)node::Buffer::Length(sbuf);
   uint64_t N = (uint64_t)Nan::To<int64_t>(info[2]).FromJust();
-  uint64_t r = (uint64_t)Nan::To<int64_t>(info[3]).FromJust();
-  uint64_t p = (uint64_t)Nan::To<int64_t>(info[4]).FromJust();
+  uint32_t r = (uint32_t)Nan::To<int64_t>(info[3]).FromJust();
+  uint32_t p = (uint32_t)Nan::To<int64_t>(info[4]).FromJust();
   size_t keylen = (size_t)Nan::To<int64_t>(info[5]).FromJust();
 
   BScryptWorker *worker = new BScryptWorker(
