+++ src/scrypt_async.cc
@@ -6,12 +6,12 @@ BScryptWorker::BScryptWorker (
   v8::Local<v8::Object> &passHandle,
   v8::Local<v8::Object> &saltHandle,
   const uint8_t *pass,
-  const uint32_t passlen,
+  const size_t passlen,
   const uint8_t *salt,
   size_t saltlen,
   uint64_t N,
-  uint64_t r,
-  uint64_t p,
+  uint32_t r,
+  uint32_t p,
   size_t keylen,
   Nan::Callback *callback
 ) : Nan::AsyncWorker(callback)
